+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * By default install with `ubuntu-session`, `yaru-theme-sound` and `yaru-theme-gnome-shell`
 * Nolonger install `MissionCenter` since it's an app image and may increase the size of the ISO file.
 * Install `gnome-system-monitor` and `gnome-sound-recorder` by default.
+* Patch Chinese localization for `shotwell.desktop`.
+* Reset arcmenu (start menu) icons to default for better localization.
 
 ## v0.0.4-alpha-jammy